pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "bonner-libraries"
description = "Assorted Python libraries developed by the Bonner Lab"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Raj Magesh Gauthaman", email = "rgautha1@jh.edu"}
]
keywords = [
    "Bonner Lab",
    "BrainIO",
    "neural datasets",
    "utilities",
    "files",
    "figshare",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
urls = {repository = "https://github.com/BonnerLab/bonner-libraries"}
dependencies = [
    "loguru",
    "tqdm",
    "more_itertools",
    "boto3",
    "numpy",
    "pandas",
    "scipy",
    "torch",
    "torchvision",
    "torchdata",
    "xarray",
    "netCDF4",
    "dask",
    "nibabel",
    "Pillow",
    "matplotlib",
    "robustbench @ git+https://github.com/RobustBench/robustbench.git@master",
    "clip @ git+https://github.com/openai/CLIP.git",
]
dynamic = ["version"]
